pyproject.toml
[tool.poetry]
name = "drf-auth"
version = "0.1.0"
description = ""
authors = ["Roman Sydoruk <sydoruk89@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.8"
django = "^3.1.1"
djangorestframework = "^3.11.1"
psycopg2-binary = "^2.8.6"
djangorestframework_simplejwt = "^4.4.0"
gunicorn = "^20.0.4"
whitenoise = "^5.2.0"
django-cors-headers = "^3.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
